move test fixtures to conftest.py
